prevent race condition for modifying files for the same preview URL
